CLN: reorg pandas/io/json to sub-dirs
