+++ pandas/io/json/__init__.py
@@ -1,0 +1,4 @@
+from .json import to_json, read_json, loads, dumps  # noqa
+from .normalize import json_normalize  # noqa
+
+del json, normalize  # noqa
