+++ setup.py
@@ -631,6 +631,7 @@ setup(name=DISTNAME,
                 'pandas.core',
                 'pandas.indexes',
                 'pandas.io',
+                'pandas.io.json',
                 'pandas.io.sas',
                 'pandas.formats',
                 'pandas.sparse',
